+++ api.finance.data.provider/src/api/core/quote/entity/Quote.java
@@ -1,0 +1,6 @@
+package api.core.quote.entity;
+
+public class Quote
+{
+
+}