+++ api.finance.data.provider.doc/changelog/changelog.md
@@ -1,9 +1,18 @@
 # API Finance Data Provider - Change Log
 
-### [1.0.0] | 2017-01-12
-
+### [1.1.0] | 2018-xx-xx
 #### ADDED
+##### General
++ API Manager with option to choose specific Finance Data Provider API
 
+##### API Simulation
++ historical quote simulation
+
+##### API Yahoo Finance
++ historical quotes
+
+### [1.0.0] | 2017-01-12
+#### ADDED
 ##### API Yahoo Finance
 + symbol lookup